COMMUNICATION-1159 - Create V2 protocol for the AM130 (IDIS P2)
Created the com.energyict.protocolimplv2.dlms.idis.am130.AM130 protocol, according to the IDIS P2 spec.
It extends the AM500 protocol, which implements the IDIS P1 spec.

git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@88747 0d108963-8a47-0410-809b-ebd9fb3b3407

+++ upl/com.energyict.mdc.upl.impl/protocolimpl/src/main/java/com/energyict/protocolimpl/dlms/idis/ObisCodeMapper.java
@@ -90,7 +90,7 @@ public class ObisCodeMapper {
             Unsigned32 value = register.getValueAttr().getUnsigned32();
             if (value != null) {
                 if (obisCode.equals(ObisCode.fromString(ALARM_REGISTER))) {
-                    AlarmBitsRegister alarmBitsRegister = new AlarmBitsRegister(obisCode, value);
+                    AlarmBitsRegister alarmBitsRegister = new AlarmBitsRegister(obisCode, value.longValue());
                     return alarmBitsRegister.getRegisterValue();
                 }
                 return new RegisterValue(obisCode, new Quantity(value.getValue(), Unit.get("")));
